+++ app/controllers/application_controller.rb
@@ -1,10 +1,10 @@
+# frozen_string_literal: true
+
 class ApplicationController < ActionController::Base
+  include Authenticable
+
   attr_reader :current_channel
 
   helper_method :current_channel
   helper_method :current_user
-
-  def current_user
-    @current_user ||= (session[:user] ||= "user-#{rand(20)}")
-  end
 end