+++ app/graphql/application_schema.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+class ApplicationSchema < GraphQL::Schema
+  default_max_page_size 50
+
+  mutation(Types::MutationType)
+  query(Types::QueryType)
+
+  # For batch-loading (see https://graphql-ruby.org/dataloader/overview.html)
+  use GraphQL::Dataloader
+end