+++ app/graphql/types/mutation_type.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module Types
+  class MutationType < Types::BaseObject
+    field :send_message,
+      mutation: Mutations::Messages::SendMessageMutation,
+      description: "Creates new message mutation"
+  end
+end