+++ config/routes.rb
@@ -1,4 +1,13 @@
+# frozen_string_literal: true
+
+require Rails.root.join("lib", "graphql_playground", "rack")
+
 Rails.application.routes.draw do
+  post "/api/graphql", to: "graphql#execute"
+  if Rails.configuration.graphql_playground_enabled
+    mount GraphQLPlayground::Rack.new, at: "/graphql"
+  end
+
   resources :channels, only: [:index, :show] do
     resources :messages, only: [:create]
   end