+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change log
 
 ## master
+- GraphQL API added. ([@fargelus][])
 - Rubocop configuration, Lefthook & Pronto. ([@fargelus][]).
 - Added docker, docker-compose, dip. ([@fargelus][]).